com.unity.test-framework.performance@2.1.0-preview ## [2.1.0] - 2020-05-14
Add flexible horizontal splitter for report window
Fix date format

## [2.0.9] - 2020-03-23

Fix profiler measurements for method measurements
Throw exceptions when measuring NaN

+++ Editor/TestRunBuilder.cs
@@ -27,7 +27,7 @@ namespace Unity.PerformanceTesting.Editor
             run.Dependencies = GetPackageDependencies();
             SetBuildSettings(run);
 
-            run.Date = (int)Utils.ConvertToUnixTimestamp(DateTime.Now);
+            run.Date = Utils.ConvertToUnixTimestamp(DateTime.Now);
 
             CreateResourcesFolder();
             CreatePerformanceTestRunJson(run);
